+++ lib/cloudinary.rb
@@ -29,8 +29,8 @@ module Cloudinary
   OLD_AKAMAI_SHARED_CDN = "cloudinary-a.akamaihd.net"
   SHARED_CDN            = AKAMAI_SHARED_CDN
 
-  USER_AGENT      = "CloudinaryRuby/" + VERSION
-  @@user_platform = ""
+  USER_AGENT      = "CloudinaryRuby/#{VERSION} (Ruby #{RUBY_VERSION}-p#{RUBY_PATCHLEVEL})"
+  @@user_platform = defined?(Rails.version) ? "Rails/#{Rails.version}" : ""
 
   # Add platform information to the USER_AGENT header
   # This is intended for platform information and not individual applications!
@@ -44,7 +44,7 @@ module Cloudinary
 
   def self.USER_AGENT
     if @@user_platform.empty?
-      "#{USER_AGENT}"
+      USER_AGENT
     else
       "#{@@user_platform} #{USER_AGENT}"
     end
@@ -86,7 +86,12 @@ module Cloudinary
 
   def self.config_from_url(url)
     @@config ||= OpenStruct.new
-    uri      = URI.parse(url)
+    return unless url && !url.empty?
+    uri = URI.parse(url)
+    if !uri.scheme || "cloudinary" != uri.scheme.downcase
+      raise(CloudinaryException,
+        "Invalid CLOUDINARY_URL scheme. Expecting to start with 'cloudinary://'")
+    end
     set_config(
       "cloud_name"          => uri.host,
       "api_key"             => uri.user,