+++ samples/user_profiles_active_storage/config/routes.rb
@@ -1,0 +1,6 @@
+Rails.application.routes.draw do
+  resources :users do
+    resources :photos, only: [:index, :show, :destroy]
+  end
+  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+end